work with TicketController

+++ java_hometasks/spring-cloud-students/ticket-service/src/main/java/com/intentsg/service/ticket/models/Ticket.java
@@ -20,7 +20,7 @@ public class Ticket {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    int ticketId;
+    long ticketId;
 
     @NotEmpty
     int hallNumber;
